ref(setup): move fixture setup to loadFixtures and improve error message
